use doubles for wav.c as we need it

+++ include/common.h
@@ -354,9 +354,9 @@ typedef struct wav
 
 typedef struct wavchn
 {
-	float offs[WAV_MAX_OUTPUTS];
-	float ldelay[WAV_MAX_OUTPUTS];
-	float lpf_charge[WAV_MAX_OUTPUTS];
+	double offs[WAV_MAX_OUTPUTS];
+	double ldelay[WAV_MAX_OUTPUTS];
+	double lpf_charge[WAV_MAX_OUTPUTS];
 	wav_t *src;
 	int idx;
 	int flags;
